+++ packages/app/src/App.tsx
@@ -92,7 +92,7 @@ const customHomePageModule = createFrontendModule({
                   <HeaderWorldClock clockConfigs={clockConfigs} />
                 </Header>
                 <Content>
-                  <CustomHomepageGrid>
+                  <CustomHomepageGrid preventDuplicateWidgets>
                     {widgets.map((widget, index) => (
                       <Fragment key={widget.name ?? index}>
                         {widget.component}