+++ packages/app/src/examples/pagesPlugin.tsx
@@ -26,6 +26,7 @@ import {
   createExtensionBlueprint,
   createExtensionInput,
   coreExtensionData,
+  BreadcrumbEntry,
 } from '@backstage/frontend-plugin-api';
 import { useEffect, useState } from 'react';
 import { Route, Routes } from 'react-router-dom';
@@ -167,8 +168,25 @@ const Page1 = PageBlueprint.make({
               Sub-page content:
               <div>
                 <Routes>
-                  <Route path="/" element={<h2>This is also page 1</h2>} />
-                  <Route path="/page2" element={<h2>This is page 2</h2>} />
+                  <Route
+                    path="/"
+                    element={
+                      <BreadcrumbEntry entry={{ label: 'Page 1', href: '/' }}>
+                        <h2>This is also page 1</h2>
+                      </BreadcrumbEntry>
+                    }
+                  />
+
+                  <Route
+                    path="/page2"
+                    element={
+                      <BreadcrumbEntry
+                        entry={{ label: 'Page 2', href: '/page2' }}
+                      >
+                        <h2>This is page 2</h2>
+                      </BreadcrumbEntry>
+                    }
+                  />
                 </Routes>
               </div>
             </div>