+++ packages/cli-module-config/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/cli-module-config
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config-loader@1.11.0
+  - @backstage/cli-common@0.3.0
+  - @backstage/cli-node@0.3.4
+
+## 0.1.4-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/cli-common@0.3.0-next.0
+  - @backstage/cli-node@0.3.4-next.0
+  - @backstage/config-loader@1.11.0-next.2
+
 ## 0.1.4-next.0
 
 ### Patch Changes