+++ packages/frontend-dev-utils/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @backstage/frontend-dev-utils
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.17.3
+  - @backstage/plugin-app@0.5.1
+  - @backstage/ui@0.17.0
+  - @backstage/frontend-defaults@0.5.4
+
 ## 0.1.4-next.0
 
 ### Patch Changes