+++ plugins/catalog-import/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/plugin-catalog-import
 
+## 0.13.15
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.17.3
+  - @backstage/plugin-catalog-react@3.2.0
+  - @backstage/core-components@0.18.12
+  - @backstage/catalog-client@1.16.1
+  - @backstage/core-plugin-api@1.12.8
+  - @backstage/integration-react@1.2.20
+  - @backstage/plugin-permission-react@0.5.3
+
 ## 0.13.15-next.0
 
 ### Patch Changes