+++ plugins/connections-example-backend-module-gitlab/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-connections-example-backend-module-gitlab
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/connections@0.2.0
+  - @backstage/backend-plugin-api@1.9.3
+
 ## 0.1.2-next.0
 
 ### Patch Changes