+++ plugins/org/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-org
 
+## 0.7.6
+
+### Patch Changes
+
+- 9585e44: Fixed the ownership card emitting duplicate React key warnings when different entity kinds share the same type.
+- Updated dependencies
+  - @backstage/frontend-plugin-api@0.17.3
+  - @backstage/ui@0.17.0
+  - @backstage/plugin-catalog-react@3.2.0
+  - @backstage/core-components@0.18.12
+  - @backstage/core-plugin-api@1.12.8
+
 ## 0.7.6-next.0
 
 ### Patch Changes