+++ plugins/permission-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-permission-backend
 
+## 0.7.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-auth-node@0.7.3
+  - @backstage/backend-plugin-api@1.9.3
+  - @backstage/plugin-permission-node@0.11.2
+
 ## 0.7.14-next.0
 
 ### Patch Changes