+++ plugins/scaffolder-backend-module-github/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-scaffolder-backend-module-github
 
+## 0.9.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.9.3
+  - @backstage/plugin-catalog-node@2.2.3
+  - @backstage/plugin-scaffolder-node@0.13.5
+
 ## 0.9.11-next.0
 
 ### Patch Changes