+++ plugins/scaffolder-node-test-utils/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @backstage/plugin-scaffolder-node-test-utils
 
+## 0.3.13
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-test-utils@1.11.5
+  - @backstage/backend-plugin-api@1.9.3
+  - @backstage/plugin-scaffolder-node@0.13.5
+
 ## 0.3.13-next.0
 
 ### Patch Changes