+++ plugins/search-backend-module-pg/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-search-backend-module-pg
 
+## 0.5.57
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.9.3
+  - @backstage/plugin-search-backend-node@1.4.6
+
 ## 0.5.57-next.0
 
 ### Patch Changes